// frontend/src/api/models/MainRepeatingQuest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Level-Up API
 * API for Level-Up app
 *
 * The version of the OpenAPI document: 0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { MainQuestsSkills } from './MainQuestsSkills';
import {
    MainQuestsSkillsFromJSON,
    MainQuestsSkillsFromJSONTyped,
    MainQuestsSkillsToJSON,
    MainQuestsSkillsToJSONTyped,
} from './MainQuestsSkills';

/**
 * 
 * @export
 * @interface MainRepeatingQuest
 */
export interface MainRepeatingQuest {
    /**
     * 
     * @type {boolean}
     * @memberof MainRepeatingQuest
     */
    completed?: boolean;
    /**
     * 
     * @type {number}
     * @memberof MainRepeatingQuest
     */
    experience?: number;
    /**
     * 
     * @type {number}
     * @memberof MainRepeatingQuest
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof MainRepeatingQuest
     */
    name?: string;
    /**
     * 
     * @type {Array<MainQuestsSkills>}
     * @memberof MainRepeatingQuest
     */
    skills?: Array<MainQuestsSkills>;
    /**
     * 
     * @type {Array<MainRepeatingQuest>}
     * @memberof MainRepeatingQuest
     */
    subQuests?: Array<MainRepeatingQuest>;
    /**
     * 
     * @type {number}
     * @memberof MainRepeatingQuest
     */
    subQuestsCompleted?: number;
    /**
     * 
     * @type {string}
     * @memberof MainRepeatingQuest
     */
    type: string;
}

/**
 * Check if a given object implements the MainRepeatingQuest interface.
 */
export function instanceOfMainRepeatingQuest(value: object): value is MainRepeatingQuest {
    if (!('type' in value) || value['type'] === undefined) return false;
    return true;
}

export function MainRepeatingQuestFromJSON(json: any): MainRepeatingQuest {
    return MainRepeatingQuestFromJSONTyped(json, false);
}

export function MainRepeatingQuestFromJSONTyped(json: any, ignoreDiscriminator: boolean): MainRepeatingQuest {
    if (json == null) {
        return json;
    }
    return {
        
        'completed': json['completed'] == null ? undefined : json['completed'],
        'experience': json['experience'] == null ? undefined : json['experience'],
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'skills': json['skills'] == null ? undefined : ((json['skills'] as Array<any>).map(MainQuestsSkillsFromJSON)),
        'subQuests': json['subQuests'] == null ? undefined : ((json['subQuests'] as Array<any>).map(MainRepeatingQuestFromJSON)),
        'subQuestsCompleted': json['subQuestsCompleted'] == null ? undefined : json['subQuestsCompleted'],
        'type': json['type'],
    };
}

export function MainRepeatingQuestToJSON(json: any): MainRepeatingQuest {
    return MainRepeatingQuestToJSONTyped(json, false);
}

export function MainRepeatingQuestToJSONTyped(value?: MainRepeatingQuest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'completed': value['completed'],
        'experience': value['experience'],
        'id': value['id'],
        'name': value['name'],
        'skills': value['skills'] == null ? undefined : ((value['skills'] as Array<any>).map(MainQuestsSkillsToJSON)),
        'subQuests': value['subQuests'] == null ? undefined : ((value['subQuests'] as Array<any>).map(MainRepeatingQuestToJSON)),
        'subQuestsCompleted': value['subQuestsCompleted'],
        'type': value['type'],
    };
}